Fixed broken reference to relocated helper files and/or contract
