+++ test/helpers/utils/sleep.js
@@ -1,7 +1,7 @@
-module.exports = {
-  sleep: async(milliseconds) => {
-    return new Promise((resolve) => {
-      setTimeout(resolve, milliseconds);
-    });
-  }
+const sleep = async(milliseconds) => {
+  return new Promise((resolve) => {
+    setTimeout(resolve, milliseconds);
+  });
 };
+
+module.exports = sleep;
