+++ test/subscriptions.js
@@ -1,7 +1,7 @@
 const Ganache = require(process.env.TEST_BUILD
   ? "../build/ganache.core." + process.env.TEST_BUILD + ".js"
   : "../index.js");
-const { promisify, generateSend } = require("./helpers/utils");
+const { promisify, generateSend } = require("./helpers/utils/rpc");
 const assert = require("assert");
 const PORT = 8545;
 const HOST = "127.0.0.1";
